+++ frontend/src/components/pages/TasksShow.tsx
@@ -1,44 +1,25 @@
-import { useContext, useEffect, useState } from "react";
-import { Link as RouterLink, useParams } from "react-router-dom";
+import { useEffect, useState } from "react";
+import { Link as RouterLink } from "react-router-dom";
 import {
   Accordion,
   AccordionDetails,
   AccordionSummary,
-  Button,
-  Card,
-  CardActions,
-  CardContent,
-  CardHeader,
-  IconButton,
   Link,
   Stack,
-  Tooltip,
   Typography,
 } from "@mui/material";
-import EditIcon from "@mui/icons-material/Edit";
-import PeopleIcon from "@mui/icons-material/People";
-import ConnectWithoutContactIcon from "@mui/icons-material/ConnectWithoutContact";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2";
 import { useFetchTask } from "../../hooks/useFetchTask";
-import { AuthContext } from "../../providers/AuthProvider";
-import { DatetimeFormat } from "../ui/DatetimeFormat";
-import { TasksDeleteIconButton } from "../models/task/TasksDeleteIconButton";
-import { UpdateIsDoneButton } from "../models/task/UpdateIsDoneButton";
 import { Task } from "../../types/taskTypes";
 import { LoadingColorRing } from "../ui/LoadingColorRing";
 import { UserNameHeader } from "../models/user/UserNameHeader";
 import { ChildrenTasksDetails } from "../models/task/ChildrenTasksDetails";
 import { ParentTaskDetails } from "../models/task/ParentTaskDetails";
-import { TaskContentTypography } from "../models/task/TaskContentTypography";
-import { PriorityStack } from "../models/task/PriorityStack";
-import { DeadlineTypography } from "../models/task/DeadlineTypography";
-import { IsDoneTypography } from "../models/task/IsDoneTypography";
+import { TaskCard } from "../models/task/TaskCard";
 import { BackIconButton } from "../ui/BackIconButton";
-import { usePatchTaskIsDone } from "../../hooks/usePatchTaskIsDone";
 
 export const TasksShow = () => {
-  const { currentUser } = useContext(AuthContext);
   const {
     task: taskData,
     user,
@@ -46,11 +27,8 @@ export const TasksShow = () => {
     childrenTasks,
     division,
   } = useFetchTask({ action: "show" });
-  const params = useParams<{ id: string }>();
 
   const [task, setTask] = useState<Task>();
-
-  const handleUpdateTaskIsDone = usePatchTaskIsDone({ task, setTask });
 
   useEffect(() => {
     if (taskData) {
@@ -77,164 +55,12 @@ export const TasksShow = () => {
               </Stack>
             </Grid2>
             <Grid2 xs={12}>
-              <Card sx={{ p: { xs: 1, sm: 2 } }}>
-                <CardHeader
-                  title={
-                    <Typography variant="h5" component="div">
-                      {task.title}
-                    </Typography>
-                  }
-                  subheader={
-                    <Grid2
-                      container
-                      direction="column"
-                      spacing={1}
-                      p={0}
-                      mt={1}
-                    >
-                      <Grid2 xs={12}>
-                        <Typography
-                          color="text.secondary"
-                          variant="body2"
-                          component="div"
-                        >
-                          登録日：{DatetimeFormat(task.created_at)}
-                        </Typography>
-                      </Grid2>
-                      {division && (
-                        <Grid2 xs={12}>
-                          <Stack direction="row">
-                            <PeopleIcon
-                              sx={{
-                                width: 18,
-                                height: 18,
-                                mr: 1,
-                              }}
-                            />
-                            <Typography
-                              color="text.secondary"
-                              variant="body2"
-                              component="div"
-                            >
-                              親タスクから分担されました
-                            </Typography>
-                          </Stack>
-                        </Grid2>
-                      )}
-                      {childrenTasks.length ? (
-                        <Grid2 xs={12}>
-                          <Stack direction="row">
-                            <PeopleIcon
-                              sx={{
-                                width: 18,
-                                height: 18,
-                                mr: 1,
-                              }}
-                            />
-                            <Typography
-                              color="text.secondary"
-                              variant="body2"
-                              component="div"
-                            >
-                              子タスクへ分担されました
-                            </Typography>
-                          </Stack>
-                        </Grid2>
-                      ) : null}
-                    </Grid2>
-                  }
-                  action={
-                    task?.user_id === currentUser?.id && (
-                      <Stack direction="row" spacing={1}>
-                        <Tooltip title="編集" placement="top" arrow>
-                          <IconButton
-                            component={RouterLink}
-                            to={`/tasks/${task?.id}/edit`}
-                          >
-                            <EditIcon />
-                          </IconButton>
-                        </Tooltip>
-                        <TasksDeleteIconButton taskId={params.id} />
-                      </Stack>
-                    )
-                  }
-                />
-                <CardContent sx={{ p: { xs: 1, sm: 2 } }}>
-                  <Grid2
-                    container
-                    rowSpacing={2}
-                    pt={0}
-                    px={0}
-                    sx={{ pb: { xs: 1, sm: 2 } }}
-                  >
-                    {task.description && (
-                      <Grid2 xs={12}>
-                        <TaskContentTypography
-                          subtitle="詳細"
-                          body={task.description}
-                        />
-                      </Grid2>
-                    )}
-                    <Grid2 xs={6}>
-                      <PriorityStack value={task.priority} />
-                    </Grid2>
-                    <Grid2 xs={6}>
-                      <DeadlineTypography deadline={task.deadline} />
-                    </Grid2>
-                    <Grid2 xs={6}>
-                      <IsDoneTypography isDone={task.is_done} />
-                    </Grid2>
-                    <Grid2 xs={6}>
-                      <TaskContentTypography
-                        subtitle="進捗率"
-                        body={`${task.rate_of_progress}%`}
-                      />
-                    </Grid2>
-                    {task.is_done && (
-                      <Grid2 xs={6}>
-                        <TaskContentTypography
-                          subtitle="完了日"
-                          body={DatetimeFormat(task.updated_at)}
-                        />
-                      </Grid2>
-                    )}
-                  </Grid2>
-                </CardContent>
-                <CardActions>
-                  <Stack direction="row" spacing={2}>
-                    {task.user_id === currentUser?.id && (
-                      <UpdateIsDoneButton
-                        onClick={handleUpdateTaskIsDone}
-                        disabled={false}
-                        isFinished={task.is_done}
-                      />
-                    )}
-                    {task.is_done || (
-                      <div>
-                        <Tooltip title="分担する" placement="top" arrow>
-                          <IconButton
-                            color="primary"
-                            component={RouterLink}
-                            to={`/tasks/${task.id}/divisions/new`}
-                            sx={{ display: { xs: "flex", md: "none" } }}
-                          >
-                            <ConnectWithoutContactIcon />
-                          </IconButton>
-                        </Tooltip>
-                        <Button
-                          variant="contained"
-                          component={RouterLink}
-                          to={`/tasks/${task.id}/divisions/new`}
-                          startIcon={<ConnectWithoutContactIcon />}
-                          sx={{ display: { xs: "none", md: "flex" } }}
-                        >
-                          分担する
-                        </Button>
-                      </div>
-                    )}
-                  </Stack>
-                </CardActions>
-              </Card>
+              <TaskCard
+                task={task}
+                setTask={setTask}
+                childrenTasks={childrenTasks}
+                division={division}
+              />
             </Grid2>
             <Grid2 xs={12}>
               {division && (
@@ -269,7 +95,7 @@ export const TasksShow = () => {
                               {division.user.name}
                             </Link>
                           ) : (
-                            "退会済みユーザー"
+                            "不明なユーザー"
                           )}
                         </Typography>
                       </Stack>
